sptool the rest
